// protocal/types/enum.ts
export enum RTTPOperation {
  ESTAB = "ESTAB",
  REGISTER_DRIVER = "REGISTER_DRIVER",
  REGISTER_PASSENGER = "REGISTER_PASSENGER",
  REQUEST_RIDE = "REQUEST_RIDE",
  OFFER_RIDE = "OFFER_RIDE",
  ACCEPT_RIDE = "ACCEPT_RIDE",
  END_RIDE = "END_RIDE",
  ERROR = "ERROR",
  NOT_IMPLEMENTED = "NOT_IMPLEMENTED",
}

export enum RTTPType {
  INFORM = "INFORM",
  ACKN = "ACKN",
}